+++ host/sgx/platformquoteprovider.h
@@ -54,6 +54,15 @@ typedef sgx_plat_error_t (*sgx_get_quote_verification_collateral_t)(
     const char* pck_ca,
     sgx_ql_qve_collateral_t** pp_qve_collateral);
 
+typedef sgx_plat_error_t (
+    *sgx_get_quote_verification_collateral_with_parameters_t)(
+    const uint8_t* fmspc,
+    const uint16_t fmspc_size,
+    const char* pck_ca,
+    const uint8_t* custom_parameters,
+    const uint16_t custom_parameters_length,
+    sgx_ql_qve_collateral_t** pp_quote_collateral);
+
 typedef void (*sgx_free_quote_verification_collateral_t)(
     sgx_ql_qve_collateral_t* p_qve_collateral);