+++ tests/attestation_plugin/plugin/tests.h
@@ -31,4 +31,8 @@ void verify_sgx_evidence(
     const uint8_t* custom_claims_buffer,
     size_t custom_claims_buffer_size);
 
+#ifndef OE_BUILD_ENCLAVE
+void set_up_mocks_for_host();
+#endif
+
 #endif // _OE_ATTESTATION_PLUGIN_TESTS